Download NPDES permits

Downloads NPDES permits from EPA ECHO API services within the bounds of the SpatRast template.

Usage

download_NPDES_permits(
  template,
  permit_component = "POT",
  permit_status = "EFF",
  output = tempfile(fileext = ".gpkg")
)

Arguments

template

A SpatRaster object.

permit_component

A character vector with one or more of the following: PRE (pretreatment), CAF (CAFO), CSO (CSO), POT (Publicly Owned Treatment Works, the default), BIO (Biosolids), SWS (Stormwater Small MS4), SWM (Stormwater Medium/Large MS4), SWI (Stormwater Industrial), SWC (Stormwater Construction).

permit_status

A character vector with one or more of the following: EFF (effective, the default), EXP (expired), PND (pending), TRM (terminated), RET (retired), NON (not needed), ADC (administratively continued).

output

A character file path specifying where the raster file should be stored. Defaults to a temporary file.

Value

SpatVector object

Examples

dem <- system.file("extdata", "thompsoncreek.tif", package = "SELECTRdata")
dem <- terra::rast(dem)
download_NPDES_permits(dem)
#> ℹ Query returned 3 results!
#>  class       : SpatVector 
#>  geometry    : points 
#>  dimensions  : 3, 23  (geometries, attributes)
#>  extent      : -96.47217, -96.41417, 30.58864, 30.66722  (xmin, xmax, ymin, ymax)
#>  source      : file20cf10283731.gpkg
#>  coord. ref. : lon/lat WGS 84 (EPSG:4326) 
#>  names       :          CWPName  SourceID        CWPStreet       CWPCity
#>  type        :            <chr>     <chr>            <chr>         <chr>
#>  values      :  RELLIS CAMPUS ~ TX0076023  APPROX 3.0 MI ~ BRAZOS COUNTY
#>                STILL CREEK WWTP TX0025071  2028 QUALITY P~         BRYAN
#>                 THOMPSONS CREE~ TX0128554 7198 VERLOREN LN         BRYAN
#>  CWPState CWPStateDistrict CWPZip MasterExternalPermitNmbr EPASystem Statute
#>     <chr>            <chr>  <chr>                    <chr>     <chr>   <chr>
#>        TX               09  77807                       NA       ICP     CWA
#>        TX               09  77805                       NA       ICP     CWA
#>        TX               09  77807                       NA       ICP     CWA
#>  (and 13 more)
#>               
#>               
#>               
#>